requirements.txt
aiohttp==3.7.4
astroid==2.4.2
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
chardet==3.0.4
colorama==0.4.4
decorator==4.4.2
discord.py==1.6.0
idna==3.1
ipykernel==5.4.3
ipython==7.16.1
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==5.1.0
parso==0.8.1
pickleshare==0.7.5
prompt-toolkit==3.0.11
Pygments==2.7.4
pylint==2.6.0
python-dateutil==2.8.1
pyzmq==21.0.1
six==1.15.0
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing-extensions==3.7.4.3
wcwidth==0.2.5
wrapt==1.12.1
yarl==1.6.3
